Updated database functionality
Created bat files for running the webscraper, webapp and database editing. Not included in this commit is the task scheduler files used for running the different bat files.

Database was updated with additional tables and a script which moves data to enable trend visualization of different stock mentions in the flask web app.